requirements.txt
python-dotenv==1.0.0
duckdb==1.0.0
datasus-db==0.1.3
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.5
streamlit==1.28.0
pandas==2.1.0
requests==2.31.0
pydantic==2.5.0
